Document `eventType` projection event envelope property.
This may not seem super useful at first glance but I've used it a few
times now (having manually inspected the envelope and discovered it).

It's particularly useful for projections with `$any` for example, we use
one like this at my work
```javascript
fromAll()
  .when({
    $init: () => ({
      eventTypes: {},
      handled: 0
    }),
    $any: (s, e) => {
      s.eventTypes[e.eventType] = (s.eventTypes[e.eventType] || 0) + 1;
      s.handled++;
    }
  });
```

Also sneak in a `javascript` language hint on projection code.